+++ contracts/finality/src/exec/finality.rs
@@ -265,6 +265,7 @@ fn ensure_fp_exists_and_not_slashed(
 
 /// `get_msg_equivocation_evidence` returns the message for an equivocation evidence.
 /// The message will be sent to Babylon for slashing
+#[allow(clippy::too_many_arguments)]
 fn get_msg_equivocation_evidence(
     env: &Env,
     fp_btc_pk: &[u8],