+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### State and API breaking
 
+* [#65](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/65) Remove `is_enabled` flag and associated functionality from finality contract
 * [#40](https://github.com/babylonlabs-io/rollup-bsn-contracts/pull/40) Timestamped public randomness commitments.
   Breaks `PubRandCommit` struct that is being used for both storage and queries.